Refactor `SuggestPort` function to be able to reuse it in `gardenadm`
