+++ pkg/utils/oci/oci_suite_test.go
@@ -16,7 +16,7 @@ import (
 	. "github.com/onsi/gomega"
 	helmregistry "helm.sh/helm/v3/pkg/registry"
 
-	"github.com/gardener/gardener/pkg/utils/test/port"
+	utilsnet "github.com/gardener/gardener/pkg/utils/net"
 )
 
 func TestOCI(t *testing.T) {
@@ -52,7 +52,7 @@ func startTestRegistry(ctx context.Context) (string, error) {
 	config := &configuration.Configuration{}
 	config.Storage = map[string]configuration.Parameters{"inmemory": map[string]interface{}{}}
 
-	port, host, err := port.SuggestPort("")
+	port, host, err := utilsnet.SuggestPort("")
 	if err != nil {
 		return "", err
 	}
